refactor: moved table generators to tables.py
