+++ src/bonbast/datadir.py
@@ -1,5 +1,6 @@
 import sys
 import pathlib
+import os
 
 from datetime import datetime
 from typing import Optional, Tuple, List
