+++ src/bonbast/main.py
@@ -14,13 +14,12 @@ try:
     from .server import get_token_from_main_page, get_prices_from_api
     from .datadir import save_token, get_token, delete_token
     from .data import Currency, Coin, Gold
+    from .tables import get_currencies_table, get_coins_table, get_gold_table
 except:
     from server import get_token_from_main_page, get_prices_from_api
     from datadir import save_token, get_token, delete_token
     from data import Currency, Coin, Gold
-
-
-PRICE_FORMATTER = "{:,}"
+    from tables import get_currencies_table, get_coins_table, get_gold_table
 
 
 BUY = '1'
@@ -50,74 +49,6 @@ def get_prices():
         return get_prices()
     else:
         return response
-
-
-def __get_currencies_sub_table(currencies: List[Currency]) -> ConsoleRenderable:
-    table = Table()
-
-    table.add_column("Code", style="cyan", no_wrap=True)
-    table.add_column("Currency", style="cyan")
-    table.add_column("Buy", style="green", no_wrap=True)
-    table.add_column("Sell", style="green", no_wrap=True)
-
-    for currency in currencies:
-        table.add_row(
-            currency.code,
-            currency.name,
-            PRICE_FORMATTER.format(currency.buy),
-            PRICE_FORMATTER.format(currency.sell),
-        )
-
-    return table
-
-
-# get currencies and prices in a table
-def get_currencies_table(currencies: List[Currency], columns: int) -> ConsoleRenderable:
-    each_part_count = math.ceil(len(currencies) / columns)
-    currencies_parts = [currencies[i:i + each_part_count] for i in range(0, len(currencies), each_part_count)]
-    tables = [__get_currencies_sub_table(part) for part in currencies_parts]
-
-    grid = Table.grid(padding=5)
-    grid.title = 'Currencies'
-    for _ in range(columns):
-        grid.add_column()
-    grid.add_row(*tables)
-
-    return grid
-
-
-# get coins and prices in a table
-def get_coins_table(coins: List[Coin]) -> ConsoleRenderable:
-    table = Table(title="Coins")
-
-    table.add_column("Coin", style="cyan")
-    table.add_column("Buy", style="green", no_wrap=True)
-    table.add_column("Sell", style="green", no_wrap=True)
-
-    for coin in coins:
-        table.add_row(
-            coin.name,
-            PRICE_FORMATTER.format(coin.buy),
-            PRICE_FORMATTER.format(coin.sell),
-        )
-
-    return table
-
-
-# get gold and it's price in a table
-def get_gold_table(golds: List[Gold]) -> ConsoleRenderable:
-    table = Table(title="Gold")
-
-    table.add_column("Gold", style="cyan")
-    table.add_column("Price", style="green", no_wrap=True)
-
-    for gold in golds:
-        table.add_row(
-            gold.name,
-            PRICE_FORMATTER.format(gold.price),
-        )
-
-    return table
 
 
 def parse_price_data(data: dict) -> Tuple[List[Currency], List[Coin], List[Gold]]:
